requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
reportlab==4.2.5
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
beautifulsoup4==4.13.3
billiard==4.2.1
black==25.1.0
blinker==1.9.0
boto3==1.38.5
botocore==1.38.5
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coverage==7.9.2
cryptography==44.0.2
daphne==4.2.1
dj-database-url==3.0.1
Django==5.1.7
django-background-tasks==1.2.8
django-redis>=5.4.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-sslserver==0.22
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
factory_boy==3.3.3
Faker==37.1.0
Flask==3.1.0
freezegun==1.5.5
h11==0.16.0
httptools==0.6.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.3
MarkupSafe==3.0.2
model-bakery==1.20.5
MouseInfo==0.1.3
msgpack==1.1.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pdfkit==1.0.0
phonenumbers==9.0.2
pillow==11.1.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.9.0
PyMsgBox==1.0.9
pyOpenSSL==25.0.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
PyYAML==6.0.2
redis==6.2.0
requests==2.32.3
s3-client==0.1.4
s3transfer==0.12.0
service-identity==24.2.0
gunicorn==21.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tabulate==0.9.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wikipedia==1.4.0
zope.interface==7.2
python-slugify>=8.0.0
redis>=5.0
django-celery-beat>=2.5.0
